+++ serialization-msgpack/src/commonMain/kotlin/com.ensarsarajcic.kotlinx.serialization.msgpack/internal/MsgPackEncoder.kt
@@ -121,6 +121,7 @@ internal class BasicMsgPackEncoder(
         if (serializer == ByteArraySerializer()) {
             encodeByteArray(value as ByteArray)
         } else {
+            println("Serializing $value with $serializer")
             super.encodeSerializableValue(serializer, value)
         }
     }